// src/services/models/communityModel/communityModel.ts
import { CommentType } from '@/components/template/commentItem/commentItem';

export type CommunityModel = {
  userID?: number;
  feedReplyID?: string;
  postDetailID?: string;
  userName?: string;
  detail?: string;
  detailHTML?: string;
  shortContent?: string;
  iFrameList?: string[];
  metaDataJson?: string;
  likeCount?: number;
  commentCount?: number;
  noofPosts?: number;
  likedByUser?: boolean;
  postDateTime?: string;
  postImageLocation?: string[];
  postDocumentsForFeed?: postDocumentsForFeed[];
  postImageLocationIDs?: string[];
  userProfileLocation?: string;
  displayAgo?: string;
  commentsReplies?: CommentsReplies[];
  expireCommentDate?: string;
  isCommentingAvailable?: boolean;
  isEditable?: boolean;
  isOwner?: boolean;
  hasNextFlag?: boolean;
  initials?: string;
  recordType?: string;
  isDeletable?: boolean;
  programId?: string;
  sessionID?: string;
  programName?: string;
  sessionName?: string;
  programTypeID?: number;
  isProgramOneOnOne?: boolean;
  showComment?: number;
  primaryFeedId?: string;
  showMore?: boolean;
  linkPreviewHtml?: string;
  embeddedIframeHtml?: string;
};
export type postDocumentsForFeed = {
  postDetailId?: string;
  creationTime?: string;
  contentType?: string;
  displayName?: string;
  location?: string;
  contentDataId?: string;
  documentTypeId?: number;
  documentName?: string;
  coverImageURL?: string;
  documentTypeName?: string;
  documentId?: string;
  contentURL?: string;
  description?: string;
  isResourceDeleted?: boolean;
  message?: string;
  progress?: string;
};

export type CommentsReplies = {
  postID?: string;
  commentID?: string;
  comment?: string;
  commentDetailHTML?: string;
  shortContent?: string;
  iFrameList?: string[];
  commentMetaDataJson?: string;
  commentLikeCount?: number;
  commentReplyCount?: number;
  commentImageLocationList?: string[];
  commentImageLocation?: string;
  userProfileLocation?: string;
  likedByUser?: boolean;
  userID?: number;
  userName?: string;
  postDateTime?: string;
  displayAgo?: string;
  expireCommentDate?: string;
  isCommentingAvailable?: boolean;
  isOwner?: boolean;
  isEditable?: boolean;
  recordType?: string;
  initials?: string;
  isDeletable?: boolean;
  primaryFeedId?: string;
  linkPreviewHtml?: string;
  embeddedIframeHtml?: string;
};
export type GetmemberlistforforumModel = {
  userProfileImage?: string;
  location?: string;
  surname?: string;
  name?: string;
  fullName?: string;
  userName?: string;
  initials?: string;
  id?: number;
  userId?: number;
  role?: string;
  isClient?: boolean;
  email?: string;
  country?: string;
  state?: string;
  city?: string;
  aboutMe?: string;
  imgIsPublic?: boolean;
};

export type FeedDetail = {
  postType?: number;
  feedReplyID?: string;
  header?: string;
  detail?: string;
  likeCount?: number;
  commentCount?: number;
  status?: number;
  isActive?: boolean;
  sessionID?: string;
  groupID?: string;
  userID?: number;
  videoDataID?: string;
  imageList?: string[];
  postImageLocationIDs?: string[];
  postImageLocation?: string[];
  PostDocumentMappingList?: string[];
  disableCommentOn?: string;
  detailHTML?: string;
  metaDataJson?: string;
  tagUserIds?: string;
  programID?: string;
  fromGlobalCalendar?: boolean;
  forUserId?: number;
  forUserName?: string;
  primaryFeedId?: string;
  id?: string;
  viewType?: CommentType;
  mainCommentId?: string;
  pageNo?: number;
};

export type GetPostDetailsForEditModel = {
  feedDetail?: FeedDetail;
  programSessionSessionName?: string;
  groupGroupName?: string;
  userName?: string;
  contentDataDisplayName?: string;
  postDocumentMappingList?: postDocumentsForFeed[];
};